DASH-WASM: availabilityTimeOffset is actually a floating number
When testing low-latency content with the DASH WASM (WebAssembly)
parser, I found out some warnings relating to the fact that
`availabilityTimeOffset`, which was here parsed as an integer, is
actually a floating point number.

It seems logical now as it is expressed in seconds, but for obscure
reason I thought previously that it was an integer.
This matters because - in both JS and Rust - you don't parse both from
a string the same way.

Though it turned out that I correctly parsed it as a float in the JS
parser, but incorrectly as a integer in the WASM parser. This fixes the
latter.