Implement better error message for the v4
Until now, the RxPlayer's errors set the Error's `name` property inside
its `message` property.

When logged in the console, such errors are automatically formatted by
the browser which generally choose to re-use the `name` to produce a
meaningful error string.
This lead to a repeat like:
`EncryptedMediaError: EncryptedMediaError (INCOMPATIBLE_KEYSYSTEMS) Some description message`

This was unnecessary and kind of ugly but we were afraid some apps were
for whatever reasons exploiting the `message` string and thus did not change
it in the v3.

For the `v4` we choose a more consize message: The code, followed by a
colon, followed by a description of what happened.
As such, the previous example would look like in a JS console when
logged:
`EncryptedMediaError: INCOMPATIBLE_KEYSYSTEMS: Some description message`

If applications want to detect the error's name/type they can
respectively rely on the more appropriate `name` or `type` property
(basically the `name` is the PascalCased name here to have an idiomatic
JS error, `type` is the RxPlayer's documented UPPER_SNAKE_CASE version
of the name).

Note that the error's message is still not part of the API and
applications should still expect that it could theoretically change its
format at any time, even if we're generally (too?) careful with that
sort of things.