Deprecate image buffer feature
In #1287, I deprecated the `NATIVE_TEXT_BUFFER` and `HTML_TEXT_BUFFER`
as they already are automatically included when importing one of the
corresponding text track parsers (so there's no point to have a separate
import here, it can just be lazily added on any of those parser's
injection to the RxPlayer).

Yet I forgot that the exact same situation was there for the
`IMAGE_BUFFER` (and even completely forgot to re-document image-related
features). This is now done.