Fix ABR on old Edge browser
We noticed a strange issue on some older Edge browser versions (like
thde one used on the XBox or some other UWP applications) where the ABR
(Adaptive BitRate, the code in charge of choosing the right audio and
video quality) would make poor quality choices, even when the user had
a high bandwidth, only in the future v4 version.

It turned out that the issue wasn't directly linked to ABR, but due to
some other assumptions wrongly assumed by the RxPlayer (only in the
future v4) which broke behavior only on older Edge browsers.

Basically:

  - The adaptive code assumed that all quality given to it were sorted
    from the lowest to the highest bitrate, as is almost convention in
    the RxPlayer code.
    Some of its algorithms were dependents on that assumption.

  - Although most of the code handling qualities respected that
    convention, the API code handling quality choices made by the
    application (or the absence of choice) broke it for an unrelated
    matter.

    Basically, to prevent sending the same quality choices consecutive
    times (purely for optimizations reasons), the API code sorted the
    array of qualities by calling the `sort` method on the array of
    objects directly.

    Not only was this totally unnecessary based on the rest of the code,
    it then used that newly sorted array to communicate it to the rest
    of the RxPlayer code and break the bitrate-sorted assumption for
    it. It turned out that Edge, surely for some cheeky reasons but in
    the right to do so :p, consistently set the first initial element
    last in the sorted array and letting the rest in order.

To fix this, I decided to stop with the bitrate-sorted assumption and to
just explicitely sort by bitrate when sorting is needed: in the ABR or
even in the API code (as sorting per bitrate is easier to think about
than sorting whole objects).