Fix LOCAL_MANIFEST feature import
We noticed that we unfortunately forgot to export the `LOCAL_MANIFEST`
feature in the last `v3.32.0` release.

It is still theoretically possible to depend on it by importing the
corresponding file buried in the `dist` directory, but it is not at
all the intended API-documented way, so this is definitely bug.

Thankfully this is not the most relied on feature but still, some
applications at Canal+ depend on it (and reported this issue to us).