Implement better exports strategy supporting both commonjs and es6-style
Keep in mind that this commit only applies for the future v4.

We noticed that various applications bundling processes expect various
ways of packaging a library, some prefer relying on commonjs-style
exports for dependencies, other on es6-style exports.

To improve on this situation, this commit implements the following way
of providing exports:

  - Both an ES2017 build with ES6 style exports and an ES5 build with
    commonJS-style exports are generated.

    The first one is compiled in `./dist/es2017`, the second one in
    `./dist/commonjs`.

  - In our package.json, we have a `main` property leading to the
    commonjs build, a `module` property leading to the es2017 one and
    an `exports` property listing all possible import paths, leading to
    one or the other, depending on properties.

I sadly had to change a little the way the RxPlayer tools are declared,
now always through a directory and inner `index.js` as the `exports`
property syntax does not seem to allow for ambiguity on whether a
path import `path.js` or `path/index.js`.

I also profited from this commit to only build the demo through
esbuild, not webpack anymore, as we didn't really rely on the
latter anymore.