Debug: update buffer graph maximum size so it becomes more readable for long contents
When displaying the RxPlayer's debug element, a graph describing the audio, video
and text qualities in the various buffers might be displayed.

At maximum, that graph could display 10000 seconds of content, which made
in that case the graph poorly readable.

I now chose to reduce it to maximum 30 minutes, as you rarely intend to see more
when displaying debug information, making it much more readable.