Demo: Add possibility to see the debug element in the demo page
The `createDebugElement` method allows to generate an RxPlayer-made debug
element.

It is used most often as a stats-for-nerds-like kind of things to
quickly inspect, debug and report issues by displaying some interesting
playback stats without having to dive into debug logs to obtain them.

We didn't yet add a possibility to show them in the demo, so this commit
quickly does that. It does it in the laziest way possible, by just
adding a "Display debug element" checkbox in the chart area of the demo,
which will display debug information on top of the media element when
enabled and playing content.

It may have more its place in the controls (which is what e.g. YouTube or
Twitch choose to do in their production page for their debug divs, and
Canal+' main products most often set it in a contextual menu), I don't
even know if it does or not, but I was too lazy to do this just for our
demo page.
Putting it in chart was much simpler because no controls with checkbox
exist yet, whereas charts only have checkboxes. So I just chose the
easiest way!