Add unsupported and undecipherable bitrates to the debug element
The debug element that might be displayed with the
`createDebugElement` method filtered from listed video and audio
bitrates those that were not for decipherable or decodable
Representation.

This makes sense, but it bothered us when debugging why some qualities
were not advertised anymore through that window on some
difficult-to-debug devices lately: was it an RxPlayer bug, a DRM policy
not respected, or codecs that were unsupported.

This commit adds audio and video bitrates which are not actually
playable, suffixing them by `E!` when it is because it's not decipherable,
by `U!` when it's because the codec is not supported, or both (`U! E!`)
when it's both.