feature: add a connection timeout (#1319)
This commit adds a new option to `loadVideo` 

`ILoadVideoOptions.requestConfig.manifest.connectionTimeout` and 
`ILoadVideoOptions.requestConfig.segment.connectionTimeout`

### Context
When performing a request, response headers are received first, then the response body is downloaded.
Downloading a response body can take time, depending on the size of the body and client bitrate, however receiving response headers should not take a long time.

You can consider that something went wrong if you did not received the response header quickly enough and you may want to abort the request, and eventually retry it.

This can be done with the new `connectionTimeout` option.  That timeout is triggered if the request hasn't received the response headers in time.

There is now two timeout options:
 - `timeout`: a global timeout triggers if the response body were not fully downloaded in time, that abort the request.
 - `connectionTimeout`: a connection timeout triggers if the response header were not received in time, that abort the request.

The configuration is independent for manifests and segments. 
The retry count is global between `timeout` and `connectionTimeout` for a request.
You should have a short `connectionTimeout` than `timeout`, otherwise this setting is useless.

### Behavior changes

- there is now a default `connectionTimeout` set to 15 000ms.
- the `timeout` option when using low-latency mode behaved like a `connectionTimeout`. This is no longer the case after this PR and it behave now as `timeout` option. Make sure the value you have set in the `timeout` option is coherent. You may want to increase this value if you previously consider it like a `connectionTimeout.`