Rename codec to codecs in the representationFilter for the v4
When porting the incoming (planned for the incoming v3.33.0) supplementalCodec
support (allowing today to better support retro-compatible Dolby Vision codecs
when playing DASH contents) to the `MULTI_THREAD` branch (#1272,
allowing to run the RxPlayer in a WebWorker optionally, even using
MSE-in-worker API when available, to improve performance and ABR
stability - this is a huge work we're pretty confident will be released,
surely in a `v4.1.0` so still not for today), we found out that the
feature was incompatible to that branch because of a subtle reason.

Basically, the supplementalCodec work on v3.x.x branches only expose as
a codec the codec that will actually be used on the device (e.g. the
Dolby Vision codec on compatible devices, or the retro-compatible
more supported codec on others). The idea is that an application should
probably only need to know which codec is actually relied on, and not
what it could have been on another device (there may be exceptions, but
it appeared to us to be the more sensible choice).

In the `MULTI_THREAD` branch however, a specific API callback which needs
to communicate that `codec` property, the `representationFilter` `loadVideo`
option, might be called on a WebWorker at a point where the codec
currently supported might not yet be known (the main browser API to check
for codec support is related to the MSE API, so it won't be available in
platforms where the MSE-in-worker feature is not available).

Here we would have three choices:
  - change the `representationFilter` API
  - make a round-trip in problematic scenarios to the main thread to
    check for codec support before calling that callback
  - choose a codec between the "supplementalCodec" one or the "codec"
    one to indicate through the `representationFilter`

We profited from the fact that the `v4.0.0` is not yet released and from
the fact that the `codec` property of a `representationFilter` does not
seem to be much used in the wild (though it's not complicated to update
the code if it is), to go with the first solution (changing the
`representationFilter` API).

The idea is here the following: `codec` was previously a simple optional
string property representing the codec of the corresponding
`Representation`. It is now an optional array of strings.

For the great majority of cases, this value will be set to either
`undefined` (meaning the codec is unknown) or to an array with a single element
which will be the actual codec relied on when the corresponding Representation
will be played.

However in the very rare scenarios, when you're:
  - playing in "multithreading mode"
  - do not have MSE-in-worker available on the current platform
  - have a `supplementalCodec` attribute and a `codec` attribute linked to the
    Representation in your MPD (for DASH only)

You'll here have an array of two codec strings, defined as a list of the
Representation's candidate codecs from the most wanted to the most compatible.
So basically in our example it would first be the Dolby Vision codec
indicated as a "supplementalCodec", then the base, more compatible, codec
indicated as a "codec" property.