Remove the necessity of installing binaryen separately
Building of the RxPlayer's WebAssembly file previously required:
  1. that the Rust compiler toolchain is installed (most likely through
     `rustup`) and that `cargo` is consequently available as a command
     globally.
  2. that the `wasm32-unknown-unknown` Rust compiler target is added
     (this is generally just basically a `rustup target add
     wasm32-unknown-unknown` call once rustup is installed).
  3. that the `binaryen` tool is installed and that `wasm-opt` is
     consequently available as a command globally.

It bothered me that we added Rust compiler toolchains requirements for
building what is a JavaScript project, even more now that we're
considering making it a requirement for the `npm run build` script.

To simplify, I try here to remove the need for the third step, which
might be the more complex to set up (`binaryen` seems to be available
through `brew` for MacOS and it is in the `extra` repository for Arch,
but I would guess that many other linux distributions wouldn't have an
easy access to it, leading to installing and updating difficulties).

To remove it, I rely on a WebAssembly build of binaryen with a JS API on
the front, through the `binaryen` npm module.
It is quite voluminous (25MB in my `node_modules` directory, making it
second after TypeScript) yet it save the need to separately have to
install and maintain a separate binary.

I then added the `script/wasm-optimize.mjs` script to interface with
that npm module and called it through the `npm run build:wasm:release`
script, which should completely replace the need for binaryen.