fix: seeking to the last frame restart the video
seeking to the last frame of the video result in restarting the video
from the beginning. This is because VideoElement.play() method restart
the play from beginning if video is ended.
According to the spec:
> If the playback has ended and the direction of playback is forwards,
> seek to the earliest possible position of the media resource.

In this case checking if the video is ended and not call play()
fixes the issue.