Fix issue arising when a track at the exact last possible position of a Period with no consecutive Period
We found out that an issue prevented from switching the track when the
current position was exactly at the position indicated by a Period's
`end` property if there was no immediately consecutive Period.

For example if a content had, as a last Period, one starting at position
`10` (seconds) and ending at `30`, and if the player was currently
paused at position `30` a `setAudioTrack` or any track switching call
wouldn't have any effect.

This is because the logic handling which Period should currently be
handled decides that a Period finishes as soon as we reached the
position indicated by its `end` property. Because under that logic
we're not playing the last Period anymore when we reached it, API
updating tracks of the current Period do not have any effect.

It actually makes perfect sense for the frequent usecase of having
consecutive Periods, where the `end` of one is equal to the `start` of
the following one (in which case the following one has priority, so
finishing the previous Period is wanted here), but it begins to show
weird behaviors like the one described here when a Period's `end` is not
shared with a consecutive Period, in which case we could (and probably
should) consider that `end` position as part of that Period instead as
there's no such ambiguity.

So this commit actually explicitely handle that case, which fixes the
issue.