[Proposal] Update src/utils so it does not depend on other directories
While trying to rely on TypeScript typechecking to ensure that code
intended to optionally run in a WebWorker aren't using any API not
available in a worker, I encountered several issues.

The easiest one to fix is with the `src/utils` directory. Even though
utils files were intended to be kind of leaves in the RxPlayer import tree
(in that they should not depend on any other code, yet may be depended on
by any other areas), I found out that TypeScript imported a high number of
files in other directories when typechecking them, especially due to types
that were declared elsewhere but also because of some `Error` types
(declared in the `src/errors` directory).

Thus I chose here to make the following modifications so that the utils
directory is truly independant of the rest of the code:
  - Errors that utils used were actually only purposed for utils. So I
    moved their declarations in the corresponding util files.

    The concerned errors were `AssertionError` (I consequently merged
    all assetion utils in `assert.ts`) and `RequestError`.
  - `initialization_segment_cache.ts` relied on very specifivc Manifest
    types. It appeared to me not generic enough to be put in `src/utils`
    so I put it closer to where it was actually used, in
    `src/core/fetchers/segment`
  - The is_node and is_worker utils were in `src/compat`.

    I guess it would make sense in both compat and utils code, yet I
    chose to put it in `src/utils` at the end to ensure `src/utils`
    didn't depend on `src/compat` at all.