Set more standard mute/unMute behaviors
Previous RxPlayer's `mute` and `unMute` method actually set the volume
to `0` instead of changing the `muted` HTMLMediaElement property.

I here propose that we update their behavior to mutate the `muted`
property instead as:
  1. It would seem more natural to an application I would guess
  2. It might integrate better with browsers built-in making use of the
     `HTMLMediaElement` properties directly (things like
     Picture-in-Picture and so on).

I also consequently updated that the `volumeChange` event to notify
about both properties, and noticed it wasn't even documented in the API
documentation.