fix: detect safari when launching from dock
safari v.17 (macOS v.14) introduce a new feature to save any webpage
as a web app. The app is added to the dock and can be used independently
from safari. Hower this app still use safari engine under the hood.
It appears that when using safari in a webapp property "window.safari"
is undefined.
As the detection of safari browser rely on this property, the browser
is mistakenly considered as non safari.
The goal of this change is to rely on userAgent that may be more
reliable.

https://support.apple.com/en-us/104996
https://developer.mozilla.org/en-US/docs/Web/HTTP/Browser_detection_using_the_user_agent#which_part_of_the_user_agent_contains_the_information_you_are_looking_for