Fix subtitles not re-displaying after disabling them
While doing our current testing for the upcoming v4.0.0-rc1 we noticed
that enabling then disabling then re-enabling subtitles lead to no
subtitles being displayed.

The issue was simple enough to fix and was linked to an optimization
where we are only running the subtitles displaying logic if there is at
least one subtitle in the buffer. Here, we forgot to reset the boolean
driving it to `false` when disabling subtitles.

Perhaps a longer term solution could be to reset it to `false` by
linking it to the `TaskCanceller` also cancelling the logic, or even
just rely on the `TaskCanceller` itself but considering we're in testing
times, I chose to do the less risky solution of having both set separately
for now.