DRM: Cleaner implementation for setMediaKeys
While doing heavy testing for the upcoming `v4.0.0-rc1` (which if
everything goes well, should be released next week), we noticed of an
issue on LG TVs linked to the fact that we were setting a `MediaKeys`
instance too close to the time when we removed it on an
`HTMLMediaElement`.

Because we were doing heavy testing and that was the first time we've
seen that issue, we decided to just put an ugly exception for LG in that
release, and to postpone a more global solution.

This is a proposal for a more global solution (so aimed to be merged for
something like a `4.0.1` release in many months potentially).

The idea is that the `setMediaKeys` typically returns a Promise, that we
were not awaiting because we've seen that on some targets, that Promise
may hang for a very long time (e.g. until segments are pushed).
Yet in the Encrypted Media Extensions (EME) W3C recommendation, it is
indicated that doing concurrent `setMediaKeys` calls may fail.
So here, we're racing with a timeout of 1 second. If it elapsed first,
we consider the MediaKeys as set.

A better solution may be to directly detect on which devices an issue
appears and on which one it does not, and returns an appropriate logic
accordingly, but this commit just propose an initial solution.

Note that this solution comes with risk: we may be adding a length of up
to 1 second to the loading time of contents on devices where
`setMediaKeys` is long. What I propose when we merge this is to identify
those devices at testing time and decide of a solution for them.